Moved geo tests into separate file
